requirements/prod.txt
Flask==1.1.4
python-dotenv==0.20.0
MarkupSafe==2.0.1
requests==2.31.0
gunicorn==20.1.0
google-cloud-storage==2.13.0